refactor(SettingFragment): 세팅 화면 리팩터링
