+++ app/src/main/java/woowacourse/movie/view/AlarmController.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.model
+package woowacourse.movie.view
 
 import android.app.AlarmManager
 import android.app.NotificationChannel
@@ -6,21 +6,22 @@ import android.app.NotificationManager
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
+import woowacourse.movie.view.model.ReservationUiModel
 import woowacourse.movie.view.seatselection.AlarmReceiver
 import java.time.ZoneId
 
-class AlarmController(private val context: Context) : AlarmManageable {
+class AlarmController(private val context: Context) {
     init {
         createChannel()
     }
 
-    override fun registerAlarms(reservations: List<ReservationUiModel>, minuteInterval: Long) {
+    fun registerAlarms(reservations: List<ReservationUiModel>, minuteInterval: Long) {
         reservations.forEach {
             registerAlarm(it, minuteInterval)
         }
     }
 
-    override fun registerAlarm(reservation: ReservationUiModel, minuteInterval: Long) {
+    fun registerAlarm(reservation: ReservationUiModel, minuteInterval: Long) {
         val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
         val pendingIntent = getPendingIntent(AlarmReceiver.newIntent(context, reservation))
 
@@ -33,13 +34,13 @@ class AlarmController(private val context: Context) : AlarmManageable {
         )
     }
 
-    override fun cancelAlarms() {
+    fun cancelAlarms() {
         val pendingIntent = getPendingIntent(Intent(context, AlarmReceiver::class.java))
         val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
         alarmManager.cancel(pendingIntent)
     }
 
-    override fun createChannel() {
+    fun createChannel() {
         val channel = NotificationChannel(AlarmReceiver.CHANNEL_ID, CHANNER_NAME, NotificationManager.IMPORTANCE_DEFAULT)
         val notificationManager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
         if (!notificationManager.notificationChannels.contains(channel)) notificationManager.createNotificationChannel(channel)
