--- app/src/main/java/woowacourse/movie/view/model/AlarmManageable.kt
@@ -1,8 +1,0 @@
-package woowacourse.movie.view.model
-
-interface AlarmManageable {
-    fun registerAlarms(reservations: List<ReservationUiModel>, minuteInterval: Long)
-    fun registerAlarm(reservation: ReservationUiModel, minuteInterval: Long)
-    fun cancelAlarms()
-    fun createChannel()
-}
