+++ app/src/main/java/woowacourse/movie/view/moviemain/setting/SettingPresenter.kt
@@ -4,13 +4,10 @@ import woowacourse.movie.data.reservation.ReservationMockRepository
 import woowacourse.movie.data.setting.SettingDataManager
 import woowacourse.movie.domain.repository.ReservationRepository
 import woowacourse.movie.view.mapper.toUiModel
-import woowacourse.movie.view.model.AlarmManageable
-import woowacourse.movie.view.moviemain.setting.SettingFragment.Companion.ALARM_MINUTE_INTERVAL
 
 class SettingPresenter(
     private val view: SettingContract.View,
     private val settingManager: SettingDataManager,
-    private val alarmController: AlarmManageable,
 ) : SettingContract.Presenter {
 
     override fun initToggle() {
@@ -21,12 +18,12 @@ class SettingPresenter(
         if (isOn && view.requestNotificationPermission()) {
             val reservationRepo: ReservationRepository = ReservationMockRepository
             val reservations = reservationRepo.findAll().map { it.toUiModel() }
-            alarmController.registerAlarms(reservations, ALARM_MINUTE_INTERVAL)
+            view.setAlarms(reservations)
             settingManager.setIsAlarmSetting(true)
             view.setToggle(true)
             return
         }
-        alarmController.cancelAlarms()
+        view.cancelAlarms()
         settingManager.setIsAlarmSetting(false)
         view.setToggle(false)
     }
